1)绑定银行卡，改为只绑定一张银行卡，不支持多张 2)查询保险产品详情，代码结构修改 3)查询用户信息，增加查询绑定银行卡信息

+++ mars-service/src/main/java/org/xteam/plus/mars/gateway/service/provider/impl/body/req/UserCardBindReqVO.java
@@ -6,18 +6,17 @@ import java.math.BigDecimal;
 public class UserCardBindReqVO implements Serializable {
 
 
-    private BigDecimal cardNo;
+    private Long cardNo;
 
     private String cardUserName;
 
     private BigDecimal bankId;
 
-
-    public BigDecimal getCardNo() {
+    public Long getCardNo() {
         return cardNo;
     }
 
-    public void setCardNo(BigDecimal cardNo) {
+    public void setCardNo(Long cardNo) {
         this.cardNo = cardNo;
     }
